apps/api/pyproject.toml
[project]
name = "keystone-api"
version = "0.0.0"
description = "Keystone API — Django + DRF backend."
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.3",
    "structlog>=25.5.0",
    "uuid-utils>=0.14.1",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
plugins = ["mypy_django_plugin.main"]
mypy_path = "."
exclude = ["^migrations/", "/migrations/"]

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["environ.*", "factory.*", "uuid_utils.*"]
ignore_missing_imports = true

# factory_boy stubs and django-stubs interactions in tests are noisy without
# adding limited value; relax the strict surface inside tests only.
[[tool.mypy.overrides]]
module = ["*.tests.*"]
disallow_untyped_calls = false
disallow_untyped_decorators = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra --strict-markers --tb=short --ds=config.settings.test"
testpaths = ["tests", "apps"]
pythonpath = ["."]

[tool.coverage.run]
branch = true
omit = ["*/migrations/*", "*/tests/*", "manage.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true

[dependency-groups]
dev = [
    "django-stubs>=6.0.3",
    "djangorestframework-stubs>=3.16.9",
    "factory-boy>=3.3.3",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.12",
]
